Move config dir to root neurolang package
