+++ neurolang/neurolang.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import, division, print_function
 
-from .utils import config
+from .config import config
 from neurolang.frontend import (
     ExplicitVBR,
     ExplicitVBROverlay,
